requirements.txt
Django==2.2.15
django-bootstrap-datepicker-plus==3.0.4
django-cleanup==4.0.0
django-contrib-comments==1.9.1
django-datatable-view==0.9.0
django-mathfilters==0.4.0
django-widget-tweaks==1.4.5
django-upload-validator==1.1.5
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
django-dbbackup-ui
html5lib==1.0.1
Pillow==6.2.1
psycopg2-binary
PyPDF2==1.26.0
python-http-client==3.2.1
pytz==2019.3
reportlab==3.5.32
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
django-multiselectfield
django-phone-field
django-phonenumber-field[phonenumbers]
django-countries
django-crispy-forms
newsapi-python
dnspython==1.16.0
boto
python-dotenv
